Disable NT5.1 jobs as they haven't worked in a while
As much as I'd love to continue supporting NT5.1 mingw job as something
that could potentially run on ReactOS, I think it's time for it to go.
The sad truth is those builds were not able to run on such systems for a
while, no matter how much we tried to keep our project not polluted with
modern symbols, our dependencies haven't been so pristine. The latest
bug with those relates to how the static executable gets created and
even after solving that issue, there comes another one that isn't ours
but a play between freetype and GCC which I'm not interested in solving.
NT5.1 is no more. It's ceased to be. NT5.1's expired and gone to meet
its maker.